requirements.txt
asgiref==3.3.1
backcall==0.2.0
basicauth==0.4.1
certifi==2020.11.8
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-basicauth==0.5.3
django-import-export==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
PyJWT==1.7.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
traitlets==4.3.3
urllib3==1.26.2
uuid==1.30
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
